project/services/analytics/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.35
pymysql==1.1.1
pymongo==4.8.0
python-dotenv==1.0.1
